refactor(community): split websocket event handlers
